Fixed the styling on New Event page, moved related files into NewEvent folder
