+++ README.md
@@ -2,8 +2,8 @@
 Group 2
 
 How to get the frontend running:
-
-    Just go into frontend directory and do npm install then npm run dev.
+0
+1    Just go into frontend directory and do npm run dev.
     Server will be running on http://localhost:3000
     There's currently just some boiler plate front end stuff from React.
