+++ frontend/src/routes/NewEvent/newevent.css
@@ -1,0 +1,50 @@
+.neweventpage{
+    display:flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+}
+
+.centered-square{
+    display: flex;
+    max-height: 80vh;
+    overflow-y: auto;
+    width: 700px;
+    border-radius: 15px;
+    box-shadow: 3px 3px 8px 3px #45b6fe;
+    background-color: white;
+    padding: 30px;
+    font-weight: bold;
+    font-size: 20px;
+    padding-bottom: 100px;
+}
+
+.pad-10{
+    padding: 10px;
+}
+
+.inputbox{
+    border-radius: 10px;
+    padding: 15px;
+    width: 600px;
+    height: 20px;
+    border-style: solid;
+}
+
+.selectbox{
+    border-radius: 10px;
+    width: 600px;
+    height: 100px;
+    border-style: solid;
+}
+
+.submit-button{
+    display: flex;
+    overflow-y: auto;
+    background-color:#0d4f85b4;
+    font-size: 16px;
+    padding: 14px 40px;
+    border-radius: 8px;
+    width: 700px;
+}
